+++ src/collective/nitf/portlets/__init__.py
@@ -1,1 +1,2 @@
 # -*- coding: utf-8 -*-
+# BBB: this module will be removed in version 3.0